Fixed the registration and sign in pages
Unified the look of the session and registration pages. 
(There is a very confusing senario going on with these pages. There are Devise and User versions of these and it is not immediately clear which are being used at first glance. This should be looked at in the future, maybe delete unused duplicate pages)