Fixed visit bug, got view time working, more bug fixes, and gems
Visit is now created if it doesn't exist in db.
Most javascript events work with respect to event tracking. The problem of a page being hidden/shown in browser or closed entirely needs to be addressed.
Active job and geocoder gems added. Active job is used for background jobs, in this case geocoding the location
Geocoder is used for location calculations and location geocoding. (I don't think Ahoy actually uses it though)
kaminari pagination bug with rails_admin fixed, thought it was fixed before but it is now.